refactor(trading): replace asset hook with pure functions
