+++ packages/suite/src/hooks/wallet/trading/form/buy/useTradingBuyFormDefaultValues.ts
@@ -8,12 +8,12 @@ import {
     type TradingBuyInfoSelector,
     type TradingCountryCode,
     buildTradingFiatOption,
+    createAssetOptionFromCryptoId,
     getDefaultCountry,
     getDefaultCountrySubdivision,
     getSupportedFiatCurrencyWithFallback,
     regional,
     selectTradingInfo,
-    useTradingAssets,
 } from '@suite-common/trading';
 import { selectBaseCurrency } from '@suite-common/wallet-core';
 
@@ -25,8 +25,7 @@ export const useTradingBuyFormDefaultValues = (
     buyInfo: TradingBuyInfoSelector | undefined,
 ): TradingBuyFormDefaultValuesProps => {
     const isTorEnabled = useSelector(selectIsTorEnabled);
-    const { coins } = useSelector(selectTradingInfo);
-    const { createAssetOptionFromCryptoId } = useTradingAssets();
+    const { coins, platforms } = useSelector(selectTradingInfo);
 
     const country = !isTorEnabled
         ? (buyInfo?.buyInfo?.country as TradingCountryCode | undefined)
@@ -39,13 +38,10 @@ export const useTradingBuyFormDefaultValues = (
         [subdivision],
     );
 
-    const defaultCrypto = useMemo(() => {
-        // coins is read via ref inside createAssetOptionFromCryptoId (stable callback);
-        // referencing it here keeps the linter active while ensuring recompute after API load.
-        void coins;
-
-        return createAssetOptionFromCryptoId(cryptoId);
-    }, [createAssetOptionFromCryptoId, cryptoId, coins]);
+    const defaultCrypto = useMemo(
+        () => createAssetOptionFromCryptoId({ coins, platforms, cryptoId }),
+        [coins, cryptoId, platforms],
+    );
 
     const baseCurrencyCode = useSelector(selectBaseCurrency);
     const suggestedFiatCurrency = getSupportedFiatCurrencyWithFallback(baseCurrencyCode);
